+++ CHANGELOG.rst
@@ -8,6 +8,9 @@ Future Release
 """""""
 - Features:
     - Add skip & skip all buttons in the dialog of the broken libraries during loading a broken session
+    - Set the default directory of the dialog during saving a state machine in accordance with the chosen library in the library tree
+    - Create the data flows & data ports automatically in the nested states
+
 
 1.0.0
 """""""